To avoid few errors: if blocks added wherever required

+++ gnowsys-ndf/gnowsys_ndf/ndf/management/commands/map_reduce.py
@@ -5,23 +5,25 @@ from django.core.management.base import BaseCommand, CommandError
 dltr=list(collection.ToReduceDocs.find({'required_for':to_reduce_doc_requirement}))
 
 for doc in dltr:
-	doc_id = doc.doc_id
-	orignal_doc = collection.Node.find_one({"_id":doc_id})
-	content_dict = dict(map_reduce(orignal_doc.content_org,mapper,reducer))
-	
-	dord = collection.ReducedDocs.find_one({"orignal_id":doc_id,'required_for':reduced_doc_requirement}) #doc of reduced docs
-	if dord:
-		dord.content=content_dict
-		dord.is_indexed = False
-		dord.save()
-	else:
-		new_doc = collection.ReducedDocs()
-		new_doc.content = content_dict
-		new_doc.orignal_id = doc_id
-		new_doc.required_for = reduced_doc_requirement
-		new_doc.is_indexed = False
-		new_doc.save()
-	doc.delete()
+	if doc:
+		doc_id = doc.doc_id
+		orignal_doc = collection.Node.find_one({"_id":doc_id})
+		if orignal_doc:
+			content_dict = dict(map_reduce(orignal_doc.content_org,mapper,reducer))
+			
+			dord = collection.ReducedDocs.find_one({"orignal_id":doc_id,'required_for':reduced_doc_requirement}) #doc of reduced docs
+			if dord:
+				dord.content=content_dict
+				dord.is_indexed = False
+				dord.save()
+			else:
+				new_doc = collection.ReducedDocs()
+				new_doc.content = content_dict
+				new_doc.orignal_id = doc_id
+				new_doc.required_for = reduced_doc_requirement
+				new_doc.is_indexed = False
+				new_doc.save()
+			doc.delete()
 	
 
 class Command(BaseCommand):
